requirements.txt
# Core Dependencies
yt-dlp>=2024.3.10
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1